Add a test for the transform model layer
